+++ src/renderer/components/deposit/add/Deposit.subcomponents.tsx
@@ -2,22 +2,22 @@ import React, { useState } from 'react'
 
 import { assetToString, baseToAsset, formatAssetAmount } from '@xchainjs/xchain-util'
 import * as FP from 'fp-ts/lib/function'
-import { useIntl } from 'react-intl'
+import { FormattedMessage, useIntl } from 'react-intl'
 
 import { Network } from '../../../../shared/api/types'
-import { AssetWithAmount } from '../../../types/asgardex'
-import { Button } from '../../uielements/button'
+import { RECOVERY_TOOL_URL } from '../../../const'
+import { AssetWithAmount1e8, AssetsWithAmount1e8 } from '../../../types/asgardex'
 import * as Styled from './Deposit.style'
 
 type AssetIconAmountProps = {
-  assetWB: AssetWithAmount
+  assetWA: AssetWithAmount1e8
   network: Network
   loading: boolean
 }
 
 const AssetIconAmount: React.FC<AssetIconAmountProps> = (props): JSX.Element => {
   const {
-    assetWB: { asset, amount },
+    assetWA: { asset, amount1e8 },
     network,
     loading
   } = props
@@ -27,8 +27,7 @@ const AssetIconAmount: React.FC<AssetIconAmountProps> = (props): JSX.Element => 
       <Styled.PendingAssetLabel asset={asset} />
       <Styled.PendingAssetAmountLabel loading={loading}>
         {formatAssetAmount({
-          amount: baseToAsset(amount),
-          decimal: amount.decimal,
+          amount: baseToAsset(amount1e8),
           trimZeros: true
         })}
       </Styled.PendingAssetAmountLabel>
@@ -38,7 +37,7 @@ const AssetIconAmount: React.FC<AssetIconAmountProps> = (props): JSX.Element => 
 
 type PendingAssetsProps = {
   network: Network
-  assets: AssetWithAmount[]
+  assets: AssetsWithAmount1e8
   loading: boolean
   onClickRecovery: FP.Lazy<void>
 }
@@ -67,18 +66,27 @@ export const PendingAssets: React.FC<PendingAssetsProps> = (props): JSX.Element 
             {assets.map((assetWB, index) => (
               <AssetIconAmount
                 network={network}
-                assetWB={assetWB}
+                assetWA={assetWB}
                 loading={loading}
                 key={`${assetToString(assetWB.asset)}-${index}`}
               />
             ))}
             <Styled.RecoveryDescription>
-              {intl.formatMessage({ id: 'deposit.add.pendingAssets.recoveryDescription' })}
+              <FormattedMessage
+                id="deposit.add.pendingAssets.recoveryDescription"
+                values={{
+                  url: (
+                    <Styled.RecoveryDescriptionLink onClick={onClickRecovery}>
+                      {RECOVERY_TOOL_URL[network]}
+                    </Styled.RecoveryDescriptionLink>
+                  )
+                }}
+              />
             </Styled.RecoveryDescription>
-            <Button onClick={onClickRecovery} typevalue="outline" color="warning">
+            <Styled.RecoveryOpenButton onClick={onClickRecovery}>
               {intl.formatMessage({ id: 'deposit.add.pendingAssets.recoveryTitle' })}
               <Styled.OpenRecoveryToolIcon />
-            </Button>
+            </Styled.RecoveryOpenButton>
           </>
         ) : (
           <></>