+++ src/renderer/contexts/WalletContext.tsx
@@ -12,6 +12,7 @@ import {
   selectedAsset$,
   loadTxs,
   getExplorerTxUrl$,
+  getExplorerAddressUrl$,
   getTxs$,
   setSelectedAsset,
   resetTxsPage
@@ -25,6 +26,7 @@ type WalletContextValue = {
   loadTxs: typeof loadTxs
   reloadBalancesByChain: typeof reloadBalancesByChain
   getExplorerTxUrl$: typeof getExplorerTxUrl$
+  getExplorerAddressUrl$: typeof getExplorerAddressUrl$
   selectedAsset$: typeof selectedAsset$
   getTxs$: typeof getTxs$
   setSelectedAsset: typeof setSelectedAsset
@@ -39,6 +41,7 @@ const initialContext: WalletContextValue = {
   balancesState$,
   chainBalances$,
   getExplorerTxUrl$,
+  getExplorerAddressUrl$,
   selectedAsset$,
   getTxs$,
   setSelectedAsset,