+++ src/renderer/services/wallet/transaction.ts
@@ -9,7 +9,7 @@ import * as BNB from '../binance'
 import * as BTC from '../bitcoin'
 import * as BCH from '../bitcoincash'
 import * as C from '../clients'
-import { ExplorerUrl$, GetExplorerTxUrl$, TxsPageLD, LoadTxsParams } from '../clients'
+import { ExplorerUrl$, GetExplorerTxUrl$, TxsPageLD, LoadTxsParams, GetExplorerAddressUrl$ } from '../clients'
 import * as ETH from '../ethereum'
 import * as LTC from '../litecoin'
 import * as THOR from '../thorchain'
@@ -20,6 +20,8 @@ import { ApiError, ErrorId, LoadTxsHandler, ResetTxsPageHandler } from './types'
 export const explorerUrl$: ExplorerUrl$ = C.explorerUrl$(client$)
 
 export const getExplorerTxUrl$: GetExplorerTxUrl$ = C.getExplorerTxUrl$(client$)
+
+export const getExplorerAddressUrl$: GetExplorerAddressUrl$ = C.getExplorerAddressUrl$(client$)
 
 /**
  * State of `LoadTxsProps`, which triggers reload of txs history