+++ src/renderer/views/wallet/AssetDetails/types.ts
@@ -17,4 +17,5 @@ export type CommonAssetDetailsProps = {
   walletAddress: O.Option<Address>
   network: Network
   historyExtraContent?: (isLoading: boolean) => React.ReactNode
+  getExplorerAddressUrl: any
 }